Grid Updates (not done, but read this)
My eclipse and git are messing up so I don't have active error checking and git won't let me push normally so I have to commit changes in the browser.  I didn't write the initial update() method in this class, so I'll need to check it out.  I'll give that a look and change it to account for patches when I get a chance (after class 4:20)